Migrate to ImGui v1.92.1 as git submodule
- Remove hardcoded ImGui files from Dependencies/
- Add official ImGui repository as git submodule in external/imgui
- Update CMakeLists.txt to use submodule ImGui sources and includes
- Fix compatibility issues with ImGui v1.92.1:
  * Fix narrowing conversion in EcoSystem.cpp
  * Replace deprecated ImGuiButtonFlags_Disabled with BeginDisabled()/EndDisabled()
  * Replace ButtonEx() calls with Button()
  * Use ImLerp() instead of custom Lerp() for ImVec4 interpolation
- Build tested successfully with latest ImGui v1.92.1

+++ Ecosystem_Project/Source/EcoSystem/EcoSystem.cpp
@@ -103,7 +103,7 @@ void Ecosystem::EcoSystem::AddCreature(Ecosystem::Creature* _c) {
 
 void Ecosystem::EcoSystem::RenderMap(void) {
   ImGui::SetNextWindowPos(ImVec2{0.f, 0.f}, ImGuiCond_Always);
-  ImGui::SetNextWindowSize(ImVec2{mnWindowX, mnWindowY}, ImGuiCond_Always);
+  ImGui::SetNextWindowSize(ImVec2{static_cast<float>(mnWindowX), static_cast<float>(mnWindowY)}, ImGuiCond_Always);
   ImGui::Begin("Simulation Space", nullptr,
                ImGuiWindowFlags_NoResize | ImGuiWindowFlags_NoCollapse |
                    ImGuiWindowFlags_NoMove |
